Add comments. Move SFT related code to a new module policy::sft.
